// typescript/interfaces.ts
export interface Reservation {
  _id: string;
  address: string;
  category_id: string;
  category_title: string;
  client_data: ClientData;
  date: string;
  gift_wrapper: boolean;
  isDeleted: boolean;
  note: string;
  package_price: number;
  package_title: string;
  ref_no: string;
  service_data: ServiceData;
  status: string;
  //   time: string;
  start_time: string;
  end_time: string;
  user_id: string;
  vendor_data: VendorData;
  createdAt: string;
}

interface ClientData {
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: "individual" | "company" | string;
  uid: string;
  _id: string;
}

interface ServiceData {
  address: string;
  city: string;
  country: string;
  description: string;
  images: string[];
  state: string;
  title: string;
  zip_code: string;
}

interface VendorData {
  firstName: string;
  lastName: string;
  profilePicture: string;
  type: "individual" | "company" | string;
  uid: string;
  _id: string;
}